# docs/openapi/assanpay-switch-v1.yaml
openapi: 3.0.3
info:
  title: Bank Al Murqarmah Integration Switch — AppInSnap API
  version: 1.0.0-draft
  description: |
    Contract-first stub for AppInSnap. Request/response bodies follow AssanPay Third-Party
    API entity shapes (see docs/AssanPay_ThirdParty_API_Docs_2.txt). Expand schemas during
    Phase 0 implementation; this file lists routes and auth for early integration planning.
servers:
  - url: http://localhost:8080
    description: Local Docker Compose
...
paths:
...
  /api/v1/account/open:
    post:
      tags: [Account]
      summary: Open Asaan Digital — returns accountNumber + IBAN (see required-fields-matrix)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Account open result
...
  /api/v1/payment/ift:
    post:
      tags: [Payment]
      summary: IFT payment — nested return
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: IftPaymentResponseEntity
  /api/v1/payment/ibft:
    post:
      tags: [Payment]
      summary: IBFT payment — flat Response_Code only
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: IbftPaymentResponseEntity
...
  /api/v1/lookups/banks:
    get:
      tags: [Lookups]
      summary: IBFT bank list (master data — do not hardcode on mobile)
      parameters:
        - name: active
          in: query
          schema: { type: boolean, default: true }
        - name: supportsIbft
          in: query
          schema: { type: boolean }
        - name: q
          in: query
          schema: { type: string }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Bank list with catalogVersion/asOf
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string }
        '304':
          description: Not modified
  /api/v1/lookups/banks/{imd}:
    get:
      tags: [Lookups]
      summary: Single bank by IMD
      parameters:
        - name: imd
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Bank detail
        '404':
          description: Unknown or inactive IMD
...
components:
  securitySchemes:
    ChannelBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Channel JWT from /api/v1/auth/token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
...
security:
  - ChannelBearer: []